+++ contracts/Proposals/Ballot/BallotInterface.sol
@@ -6,6 +6,6 @@ interface BallotInterface {
     function getLabel(uint index) external view returns (bytes32);
     function getData(uint index) external view returns (bytes32);
     function isValidChoice(uint index) external view returns (bool);
-    function optionsLength() external view returns (uint256);
+    function optionsLength() external view returns (uint);
 
 }
